Transpositiotaulun aloitus

+++ Tiralabra_maven/src/test/java/tiralabra/game/BoardTest.java
@@ -5,12 +5,13 @@
  */
 package tiralabra.game;
 
+import java.math.BigInteger;
 import org.junit.After;
 import org.junit.AfterClass;
+import static org.junit.Assert.*;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import static org.junit.Assert.*;
 import tiralabra.utilities.BoardUtilities;
 import tiralabra.utilities.ZobristHash;
 
@@ -157,7 +158,7 @@ public class BoardTest {
         board.setBoard( BoardUtilities.createPlayerTable(diagonalTestBoard1), Player.WHITE);
 
         ZobristHash hasher = new ZobristHash(4, 4);
-        long originalHash = hasher.hash(board.getBoard());
+        BigInteger originalHash = hasher.hash(board.getBoard());
         board.placeTile(0, 0);
         board.undo();
         assertEquals(originalHash, hasher.hash(board.getBoard()));
@@ -181,7 +182,7 @@ public class BoardTest {
         board.placeTile(3, 5);
         
         ZobristHash hasher = new ZobristHash(8, 8);
-        long hash = hasher.hash(board.getBoard());
+        BigInteger hash = hasher.hash(board.getBoard());
         assertEquals(hash, hasher.hash(board.getBoard()));
 
         board.pass();
